Theme style refactoring
